fn_8003ADC4: drop redundant (f64) round-trip in the clamp bound; plain (s32) of the f32 (94.25->94.54, unit 99.212->99.224)

+++ src/main/objprint.c
@@ -1198,9 +1198,9 @@ void fn_8003ADC4(GameObject* obj, void* tgt, void* p3, int a, u8 inv, int b)
                 }
                 else
                 {
-                    if (v > (s16)(int)(f64)prodB)
+                    if (v > (s16)(s32)prodB)
                     {
-                        v = (int)(f64)prodB;
+                        v = (s32)prodB;
                     }
                     w = (s16)v;
                 }
